Subscription destruction unit tests. and future functionality unit tests (sub object update/removal)

+++ sync/services/sync.service.js
@@ -196,7 +196,7 @@ function syncProvider() {
          */
 
         function Subscription(publication, scope) {
-            var timestampField, isSyncingOn = false,
+            var timestampField, isSyncingOn = false, destroyed,
                 isSingleObjectCache, updateDataStorage, cache, isInitialPushCompleted, deferredInitialization;
             var onReadyOff, formatRecord;
             var reconnectOff, publicationListenerOff, destroyOff;
@@ -213,6 +213,7 @@ function syncProvider() {
             var syncListener = new SyncListener();
 
             //  ----public----
+            this.toString = toString;
             this.getPublication = getPublication;
             this.getIdb = getId;
             this.ready = false;
@@ -268,18 +269,30 @@ function syncProvider() {
             function getId() {
                 return subscriptionId;
             }
-            
+
             function getPublication() {
                 return publication;
             }
 
+            function toString() {
+                return publication + ' ' + JSON.stringify(subParams);
+            }
             /**
              * destroy this subscription but also dependent subscriptions if any
              */
             function destroy() {
+                if (destroyed) {
+                    return;
+                }
+                destroyed = true;
+                if (thisSub.parentSubscription) {
+                    logDebug('Destroying Sub Subscription to ' + thisSub + ', parentSubscription to ' + thisSub.parentSubscription);
+                } else {
+                    logDebug('Destroying Subscription to ' + thisSub);
+                }
                 syncOff();
-                logDebug('Destroy subscription to ' + publication);
                 destroyDependentSubscriptions();
+                logDebug('Subscription to ' + thisSub + ' destroyed.');
             }
 
             function destroyDependentSubscriptions() {
@@ -291,10 +304,10 @@ function syncProvider() {
                     deps.push(sub.getPublication());
                     sub.destroy();
                 });
-                if (deps.length > 0) {
-                    deps = _.uniq(deps);
-                    logDebug('Destroy its ' + allSubscriptions.length + ' dependent subscription(s)  [' + deps + ']');
-                }
+                // if (deps.length > 0) {
+                //     deps = _.uniq(deps);
+                //     logDebug('Destroy its ' + allSubscriptions.length + ' dependent subscription(s)  [' + deps + ']');
+                // }
             }
 
             /** this will be called when data is available 
@@ -1032,7 +1045,10 @@ function syncProvider() {
                 if (subscriptionId) {
                     $socketio.fetch('sync.unsubscribe', {
                         version: SYNC_VERSION,
-                        id: subscriptionId
+                        id: subscriptionId,
+                        // following only useful for unit testing
+                        publication:publication,
+                        params:subParams
                     });
                     subscriptionId = null;
                 }
